requirements.txt
Appium-Python-Client==2.11.1
pytest==7.4.4
selenium==4.11.2
wait-for==1.2.0
